refact: ArchiveDetailView 서브뷰로 분리
시간 포매팅 함수 DataUtils로 이동
